New Feature: integrating TinyMCE as rich text solution for item description fields.
Comes with system settings to configure the most important aspects of the RTE (custom buttons, plugins, theme, size) and a switch to turn it on/off which defaults to off. (perhaps enable by default? Not sure on the implications of such an "in your face" release of a new feature, especially when you don't use the description field for lengthy text.
(cherry picked from commit b68015831c63899c43e4d45e5b93f8780ecf88c4)