Added to renderCssJs method. This now allows for the passing in of a new param from the tag called "gallerifficRenderJsOnStartup".
If not set all works as it did when rendering Js to screen using regClientStartupScript and regClientStartupHTMLScript. If set to zero (&gallerifficRenderJsOnStartup=`0`) then Js is rendered using regClientScript and regHTMLBlock.

This allows for users whos site run all js at the end of the page to move the default position of the js to before </body> rather then before </head>