Modx 2.3 Compatibility Fix
The album page was not working properly in 2.3.  These changes fix the
errors.  ItemSelector and listeners added.  There were new css classes
in the 2.3 media manager, so they were added here, but the old classes
retained for backwards compatibility (modx-pb-thumb css from 2.2.14 is
not needed with these classes added).  PhpThumb was not working in 2.3,
so 2.3.1 is needed for this to work.  The controller file seems to be
working in the front and back end without changes now that this js file
has been updated.  The only thing I could not find was the way to change
the default height for gallery.backend_thumb_height through the code.  I
went into system settings and manually updated this to 80 for 2.3.
Again, not sure what the proper way is to include that change through
the package manager updates.